pyproject.toml
[tool.poetry]
name = "pubmed-pharma-papers"
version = "0.1.0"
description = "A tool to fetch research papers from PubMed with authors affiliated with pharmaceutical companies"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "pubmed_pharma_papers"}]

[tool.poetry.dependencies]
python = "^3.9"
biopython = "^1.81"
requests = "^2.31.0"
rich = "^13.4.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
mypy = "^1.4.1"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
get-papers-list = "pubmed_pharma_papers.cli:main"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true